feat(kong.var) add set/get_by_index and set/get_by_name functions (#20)
### Summary

The flamegraphs of Kong 2.6.0 still show Nginx variables taking
a lot of time to read and set. This is partly because of the metatable
magic used to read and set them. This PR adds a couple of functions
to `kong.var` that I look forward to test how they improve the Kong
flamegraphs further.