chore(*) add a couple of functions for better macOS+M1 support
### Summary

There is a LuaJIT bug that is not yet resolved and it causes
issues with function calls on macOS+M1. Apple has decided to
break ABI compatibility with ARM64, and thus here we workaround
the issue described here:

https://github.com/LuaJIT/LuaJIT/issues/205

Proper fix is to fix LuaJIT, but as that issue has been open since
2016, I thought it would probably be faster to fix this with couple
of forward declarations.

This also needs changes in `resty.core.shdict` and `resty.core.response`
to take advantage of these new functions. New functions just take struct
in and forward the call on C-side to original functions. The issue arises
with macOS+M1 if functions have around 8+ arguments.