chore(*): add check of get_request() and the missing ';' (#51)
Reason of adding this check: Although it's indeed redundent given our phase
check. but it gives us extra protection because this a private API from
openresty, it doesn't guaranteeded to be unchanged in the future.
And also given the severity when it happens, it will likely cause a segfault,
instead of a lua vm crach.