fix(vars): use strong random value to construct request id
We found dup kong_request_id in customer environment. The reason for the dup
request id is caused by the random seed in nginx:

srandom(((unsigned) ngx_pid << 16) ^ tp->sec ^ tp->msec);

The seed is not safe if we run kong in different pods. Because the ngx_pid
might be the same. If so, the entrophy from the seed might not be enough.

Changed to use RAND_bytes from openssl for generating the request id.